create patterns module
